프론트 컴포넌트 및 파일 구조 완성, 에러해결, 서버 실행 테스트 완료
