+++ SsedamPet_front/src/pages/home/styles.js
